moved debug interceptors to different file
